Moved folders with indexed data to the node folder
