+++ back_end/node/indexed_data/voting_dto.go
