+++ back_end/node/node.go
@@ -1,20 +1,20 @@
 package node
 
 import (
-	"digital-voting/account_manager"
-	"digital-voting/indexed_data"
+	"digital-voting/node/account_manager"
+	indexed_data2 "digital-voting/node/indexed_data"
 )
 
 type Node struct {
 	AccountManager *account_manager.AccountManager
-	GroupProvider  *indexed_data.GroupProvider
-	VotingProvider *indexed_data.VotingProvider
+	GroupProvider  *indexed_data2.GroupProvider
+	VotingProvider *indexed_data2.VotingProvider
 }
 
 func NewNode() *Node {
 	return &Node{
 		AccountManager: account_manager.NewAccountManager(),
-		GroupProvider:  indexed_data.NewGroupProvider(),
-		VotingProvider: indexed_data.NewVotingProvider(),
+		GroupProvider:  indexed_data2.NewGroupProvider(),
+		VotingProvider: indexed_data2.NewVotingProvider(),
 	}
 }
