+++ back_end/transaction/transaction_specific/tx_account_creation.go
@@ -3,8 +3,8 @@ package transaction_specific
 import (
 	"crypto/sha256"
 	"digital-voting/account"
-	"digital-voting/account_manager"
 	"digital-voting/node"
+	"digital-voting/node/account_manager"
 	"digital-voting/signature/keys"
 	"encoding/base64"
 	"encoding/json"
